+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 3. ARM_FE - Designing and Implementing Student Event Search by Date and Roll No.
 4. ARM_FE - Designing and Implementing Alumni Details.
 5. ARM_FE - Designing and Implementing Attendance History.
+6. ARM_FE - Implementing Move To Alumni Details in Student Info Page.